add support for 3rdparty jvm deps
This roughly follows the pants approach -- inferring provided
packages from maven coordinates and falling back to a hardcoded
override map as needed. It's unclear if we'll end up using any
of this given Apple's gazelle support for rules_jvm just landed,
but was an interesting proof of concept. It's also possible the
Apple approach is slow enough that we prefer a more naive
implementation anyway.